Added a new lambda for chat deletion
delete_chats is a lambda that implements delete_batch from
expenses_persistence repository implementation. It uses a bunch of
chat_ids that are retrieved by user get_by function from the same
package and then uses only the chat_id field of each chat and sends it
to the delete_batch function.